+++ spec/spec_helper.rb
@@ -4,8 +4,6 @@ require 'rubocop-rake'
 require 'rubocop/rspec/support'
 
 RSpec.configure do |config|
-  config.include RuboCop::RSpec::ExpectOffense
-
   config.disable_monkey_patching!
   config.raise_errors_for_deprecations!
   config.raise_on_warning = true